fix: move frontend out of lib directory
Currently our frontend is built and moved to the lib/build directory in the package. Then the
package is built and the files are compiled to the lib directory. This overwrites the frontend. Now
the frontend lives in the build directory and the other files live in the lib directory.